feat(modularization): S-9 -- extract connectors/vault frontend to web/modules/connectors.js
Moves the full connectors + vault UI section (~2025 lines) from the monolithic
web/app.js into web/modules/connectors.js as an ES module.

Exports: initConnectors, loadConnectors, loadVaultPage
DI pattern: openModal/closeModal injected via initConnectors({...})
Local helpers: mainAgentId(), escapeHtml() (inlined)
Imports: showToast (toast.js), t (i18n.js)

Covers: connector-list/cards, add/configure/delete UI, connectors.hu
catalog/install-flow, SSH keys/servers, vault page, status display.

app.js: added import + initConnectors({ openModal, closeModal }) call;
connectors block removed (app.js: 9354 -> 7330 lines).

Verified: node --check, npm run build, 2873 vitest tests passing,
Playwright headless smoke 0 JS errors (connectors/memories/agents/kanban pages).
